requirements.txt
faiss-cpu==1.9.0.post1
sentence-transformers==3.3.1
google-generativeai==0.8.3
gradio==5.9.1
pypdf==5.1.0
numpy>=1.26.0,<2.0.0
pydantic>=2.0.0
python-dotenv==1.0.1
